requirements.txt
mss
Pillow